+++ scripts/uni_00_x=delta_y=eta_v=S_Pearson.py
@@ -2,12 +2,13 @@
 import os 
 import sys
 
+# qom modules
+from qom.utils.wrappers import wrap_looper
+
 # add path to local libraries
 sys.path.append(os.path.abspath(os.path.join('..', 'sync_bi_uni')))
 # import system
 from systems.Uni00 import Uni00
-# import wrapper
-from utils.wrappers import wrap_looper
 
 # all parameters
 params = {
@@ -68,4 +69,5 @@ params = {
 }
 
 # get Pearson synchronization
-looper = wrap_looper(Uni00, params, 'measure_average', 'XYLooper', 'H:/Workspace/VSCode/Python/sync_bi_uni/data/uni_00/S_Pearson_cyc_10_nth_0', True)
+looper = wrap_looper(Uni00, params, 'measure_average', 'XYLooper', 'H:/Workspace/VSCode/Python/sync_bi_uni/data/uni_00/S_Pearson_1e4-20pi', True)
+print(looper.get_thresholds(thres_mode='minmax'))