+++ systems/Uni01.py
@@ -5,7 +5,7 @@
 
 __authors__ = ['Sampreet Kalita']
 __created__ = '2020-01-28'
-__updated__ = '2021-05-12'
+__updated__ = '2021-07-04'
 
 # dependencies
 import numpy as np
@@ -44,17 +44,20 @@ class Uni01(DODMSystem):
             'n_ths': params.get('n_ths', [0, 0]),
             'omega_m': params.get('omega_m', 1.0),
         }
-        # matrices
+        # drift matrix
         self.A = None
-        self.D = None
 
-    def get_A(self, modes):
+    def get_A(self, modes, params, t):
         """Function to obtain the drift matrix.
 
         Parameters
         ----------
         modes : list
-            Values of the optical and mechancial modes.
+            Values of the modes.
+        params : list
+            Constant parameters.
+        t : float
+            Time at which the rates are calculated.
         
         Returns
         -------
@@ -63,20 +66,21 @@ class Uni01(DODMSystem):
         """
 
         # extract frequently used variables
-        eta     = self.params['eta']
-        g_0s    = self.params['g_0s']
-        gammas  = self.params['gammas']
-        kappas  = self.params['kappas']
-        temp = np.sqrt(eta * kappas[0] * kappas[1])
+        Delta_0s= [params[1], params[2]]
+        eta     = params[3]
+        g_0s    = [params[4], params[5]]
+        gammas  = [params[6], params[7]]
+        kappas  = [params[8], params[9]]
+        omega_ms= [params[10], params[11]]
         alphas  = [modes[0], modes[2]]
         betas   = [modes[1], modes[3]]
+        temp    = np.sqrt(eta * kappas[0] * kappas[1])
 
-        # get frequencies
-        omega_ms, Delta_0s = self.get_frequencies()
-
-        # effective detunings
+        # initialize lists
         Deltas  = list()
         gs      = list()
+
+        # effective detunings
         for i in range(2):
             Deltas.append(Delta_0s[i] + 2 * g_0s[i] * np.real(betas[i]))
             gs.append(g_0s[i] * alphas[i])
@@ -118,77 +122,113 @@ class Uni01(DODMSystem):
 
         return self.A
 
-    def get_D(self):
-        """Function to obtain the noise correlation matrix.
+    def get_ivc(self):
+        """Function to obtain the initial values and constants required for the IVP.
         
         Returns
         -------
-        D : list
-            Noise correlation matrix.
+        iv : list
+            Initial values of variables.
+            First element contains the optical mode of first cavity.
+            Next element contains the mechanical mode of first cavity.
+            Next element contains the optical mode of second cavity.
+            Next element contains the mechanical mode of second cavity.
+            Next (4*2)^2 elements contain the correlations.
+
+        c : list
+            Constant parameters.
+            First (4*2)^2 elements contain the noise matrix.
+            Next element contains the laser amplitude.
+            Next 2 elements contain the laser detunings.
+            Next element contains the transmission loss.
+            Next 2 elements contain the optomechanical coupling strengths.
+            Next 2 elements contain the mechanical decay rates.
+            Next 2 elements contain the optical decay rates.
+            Next 2 elements contain the mechanical frequencies.
         """
 
         # extract frequently used variables
+        A_l     = self.params['A_l']
+        Delta_0 = self.params['Delta_0']
+        delta   = self.params['delta']
         eta     = self.params['eta']
+        g_0s    = self.params['g_0s']
         gammas  = self.params['gammas']
         kappas  = self.params['kappas']
         n_ths   = self.params['n_ths']
+        omega_m = self.params['omega_m']
         temp    = np.sqrt(eta * kappas[0] * kappas[1])
+ 
+        # initial mode values as 1D list
+        modes_0 = np.zeros(4, dtype=np.complex_).tolist()
+
+        # initial quadrature correlations
+        corrs_0 = np.zeros([8, 8], dtype=np.float_)
+        for i in range(2):
+            # alternate index 
+            _ai = 1 if i == 0 else 0
+            corrs_0[4*i + 0][4*i + 0] = 0.5
+            corrs_0[4*i + 0][4*_ai + 0] = 0.5
+            corrs_0[4*i + 1][4*i + 1] = 0.5
+            corrs_0[4*i + 1][4*_ai + 1] = 0.5
+            corrs_0[4*i + 2][4*i + 2] = (n_ths[i] + 0.5)
+            corrs_0[4*i + 2][4*_ai + 2] = (n_ths[i] + 0.5)
+            corrs_0[4*i + 3][4*i + 3] = (n_ths[i] + 0.5)
+            corrs_0[4*i + 3][4*_ai + 3] = (n_ths[i] + 0.5)
+
+        # convert to 1D list and concatenate all variables
+        iv = modes_0 + [np.complex_(element) for element in corrs_0.flatten()]
 
         # noise correlation matrix
-        if self.D is None or np.shape(self.D) != (8, 8):
-            self.D = np.zeros([8, 8], dtype=np.float_)
+        D = np.zeros([8, 8], dtype=np.float_)
         for i in range(2):
             # sign of mode
             _sign = - 2 * (i - 0.5)
             # alternate index 
             _ai = 1 if i == 0 else 0
             # X
-            self.D[4*i + 0][4*i + 0] = kappas[0] / 2 + kappas[1] / 2 + _sign * temp
-            self.D[4*i + 0][4*_ai + 0] = kappas[0] / 2 - kappas[1] / 2
+            D[4*i + 0][4*i + 0] = kappas[0] / 2 + kappas[1] / 2 + _sign * temp
+            D[4*i + 0][4*_ai + 0] = kappas[0] / 2 - kappas[1] / 2
             # Y
-            self.D[4*i + 1][4*i + 1] = kappas[0] / 2 + kappas[1] / 2 + _sign * temp
-            self.D[4*i + 1][4*_ai + 1] = kappas[0] / 2 - kappas[1] / 2
+            D[4*i + 1][4*i + 1] = kappas[0] / 2 + kappas[1] / 2 + _sign * temp
+            D[4*i + 1][4*_ai + 1] = kappas[0] / 2 - kappas[1] / 2
             # Q
-            self.D[4*i + 2][4*i + 2] = gammas[0] * (n_ths[0] + 1 / 2) + gammas[1] * (n_ths[1] + 1 / 2)
-            self.D[4*i + 2][4*_ai + 2] = gammas[0] * (n_ths[0] + 1 / 2) + gammas[1] * (n_ths[1] + 1 / 2)
+            D[4*i + 2][4*i + 2] = gammas[0] * (n_ths[0] + 0.5) + gammas[1] * (n_ths[1] + 0.5)
+            D[4*i + 2][4*_ai + 2] = gammas[0] * (n_ths[0] + 0.5) + gammas[1] * (n_ths[1] + 0.5)
             # P
-            self.D[4*i + 3][4*i + 3] = gammas[0] * (n_ths[0] + 1 / 2) + gammas[1] * (n_ths[1] + 1 / 2)
-            self.D[4*i + 3][4*_ai + 3] = gammas[0] * (n_ths[0] + 1 / 2) + gammas[1] * (n_ths[1] + 1 / 2)
-
-        return self.D
-
-    def get_frequencies(self):
-        """Function to obtain the mechanical frequencies and laser detunings.
-        
-        Returns
-        -------
-        omega_ms : list
-            Mechanical frequencies.
-        Delta_0s : list
-            Laser detunings.
-        """
-        
-        # extract frequently used variables
-        delta       = self.params['delta']
-        Delta_0     = self.params['Delta_0']
-        omega_m     = self.params['omega_m']
+            D[4*i + 3][4*i + 3] = gammas[0] * (n_ths[0] + 0.5) + gammas[1] * (n_ths[1] + 0.5)
+            D[4*i + 3][4*_ai + 3] = gammas[0] * (n_ths[0] + 0.5) + gammas[1] * (n_ths[1] + 0.5)
 
         # mechanical frequencies
         omega_ms = [omega_m, omega_m + delta]
         # laser detunings
         Delta_0s = omega_ms
-        if Delta_0 < 0:
-            Delta_0s = - 1 * omega_ms
-            
-        return omega_ms, Delta_0s
-
-    def get_rates_modes(self, values):
-        """Function to obtain the rates of the optical and mechanical modes.
         
+        # constant parameters
+        params = [A_l] + \
+            Delta_0s + \
+            [eta] + \
+            g_0s + \
+            gammas + \
+            kappas + \
+            omega_ms
+
+        # all constants
+        c = D.flatten().tolist() + params
+
+        return iv, c
+
+    def get_mode_rates(self, modes, params, t):
+        """Function to obtain the rates of the optical and mechanical modes.
+
         Parameters
         ----------
-        values : list
+        modes : list
             Values of the modes.
+        params : list
+            Constants parameters.
+        t : float
+            Time at which the rates are calculated.
         
         Returns
         -------
@@ -197,16 +237,15 @@ class Uni01(DODMSystem):
         """
         
         # extract frequently used variables
-        A_l     = self.params['A_l']
-        eta     = self.params['eta']
-        g_0s    = self.params['g_0s']
-        gammas  = self.params['gammas']
-        kappas  = self.params['kappas']
-        alphas  = [values[0], values[2]]
-        betas   = [values[1], values[3]]
-
-        # get frequencies
-        Delta_0s, omega_ms = self.get_frequencies()
+        A_l     = params[0]
+        Delta_0s= [params[1], params[2]]
+        eta     = params[3]
+        g_0s    = [params[4], params[5]]
+        gammas  = [params[6], params[7]]
+        kappas  = [params[8], params[9]]
+        omega_ms= [params[10], params[11]]
+        alphas  = [modes[0], modes[2]]
+        betas   = [modes[1], modes[3]]
 
         # initialize lists
         Deltas      = list()
@@ -230,92 +269,3 @@ class Uni01(DODMSystem):
         mode_rates = [dalpha_dts[0], dbeta_dts[0], dalpha_dts[1], dbeta_dts[1]]
 
         return mode_rates
-
-    def ivc_func(self):
-        """Function to obtain the initial values and constants required for the IVP.
-        
-        Returns
-        -------
-        iv : list
-            Initial values of variables.
-
-        c : list
-            Constant parameters.
-        """
-
-        # extract frequently used variables
-        # mechanical mode frequency
-        n_ths    = self.params['n_ths']
- 
-        # initial mode values as 1D list
-        u_0 = np.zeros(4, dtype=np.complex_).tolist()
-
-        # initial quadrature correlations
-        V_0 = np.zeros([8, 8], dtype=np.float_)
-        for i in range(2):
-            # alternate index 
-            _ai = 1 if i == 0 else 0
-            V_0[4*i + 0][4*i + 0] = 1 / 2
-            V_0[4*i + 0][4*_ai + 0] = 1 / 2
-            V_0[4*i + 1][4*i + 1] = 1 / 2
-            V_0[4*i + 1][4*_ai + 1] = 1 / 2
-            V_0[4*i + 2][4*i + 2] = (n_ths[i] + 1 / 2)
-            V_0[4*i + 2][4*_ai + 2] = (n_ths[i] + 1 / 2)
-            V_0[4*i + 3][4*i + 3] = (n_ths[i] + 1 / 2)
-            V_0[4*i + 3][4*_ai + 3] = (n_ths[i] + 1 / 2)
-
-        # convert to 1D list and concatenate all variables
-        v = u_0 + [np.complex_(element) for element in V_0.flatten()]
-
-        # return concatenated lists of variables and constants
-        return v, None
-
-    def ode_func(self, t, v):
-        """Model function for the rate equations of the modes and correlations.
-        
-        The variables are complex-valued, hence the model requires a complex-valued integrator.
-        
-        Parameters
-        ----------
-        t : *float*
-            Time at which the rate is calculated.
-
-        v : list
-            Complex-valued variables defining the system. 
-            First 2 elements contain the optical and mechanical modes of the 1st cavity.
-            Next 2 elements contain the optical and mechanical modes of the 2nd cavity.
-            Next (4*2)^2 elements contain the correlations.
-
-        Returns
-        -------
-        rates : list
-            Rates of the complex-valued variables defining the system. 
-            First 2 elements contain the optical and mechanical modes of the 1st cavity.
-            Next 2 elements contain the optical and mechanical modes of the 2nd cavity.
-            Next (4*2)^2 elements contain the correlations.
-        """
-
-        # extract modes and correlations
-        modes   = v[0:4]
-        corrs   = np.real(v[4:]).reshape([8, 8])
-
-        # mode rates
-        mode_rates  = self.get_rates_modes(modes)
-
-        # drift matrix
-        A = self.get_A(modes)
-        # noise matrix
-        D = self.get_D()
-
-        # quadrature correlation rate equation
-        dcorrs_dt = A.dot(corrs) + corrs.dot(np.transpose(A)) + D
-
-        # mirror matrix
-        for i in range(len(dcorrs_dt)):
-            for j in range(0, i):
-                dcorrs_dt[i, j] = dcorrs_dt[j, i]
-
-        # convert to 1D list and concatenate all rates
-        rates = mode_rates + [np.complex(element) for element in dcorrs_dt.flatten()]
-
-        return rates